requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
coverage==6.2
dj-database-url==0.5.0
Django==4.0.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==11.3.0
faker-vehicle==0.2.0
flake8==4.0.1
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-network==0.0.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
whitenoise==5.3.0
